Re-add asset splitting
