+++ assets/server/login/change-password.html
@@ -74,13 +74,8 @@
         btnSubmit.disabled = !checkPasswordValid(inputNewPassword.value, inputRetype.value, requirements);
       }
 
-      inputNewPassword.addEventListener('change', disableSubmitIfInvalid);
-      inputNewPassword.addEventListener('keyup', disableSubmitIfInvalid);
-      inputNewPassword.addEventListener('paste', disableSubmitIfInvalid);
-
-      inputRetype.addEventListener('change', disableSubmitIfInvalid);
-      inputRetype.addEventListener('keyup', disableSubmitIfInvalid);
-      inputRetype.addEventListener('paste', disableSubmitIfInvalid);
+      addEventListeners(inputNewPassword, 'change keyup paste', disableSubmitIfInvalid);
+      addEventListeners(inputRetype, 'change keyup paste', disableSubmitIfInvalid);
 
       checkPasswordValid('', '', requirements);
