+++ assets/server/realmadmin/_form_abuse_prevention.html
@@ -113,9 +113,7 @@
       }
     }
 
-    inputLimitFactor.addEventListener('change', recalculate);
-    inputLimitFactor.addEventListener('keyup', recalculate);
-    inputLimitFactor.addEventListener('paste', recalculate);
+    addEventListeners(inputLimitFactor, 'change keyup paste', recalculate);
   });
 </script>
