+++ assets/server/realmadmin/_stats_users.html
@@ -69,9 +69,7 @@
         return;
       }
 
-      while(containerChart.firstChild) {
-        containerChart.removeChild(containerChart.firstChild);
-      }
+      clearChildren(containerChart)
 
       let containerListGroup = document.createElement('div');
       containerListGroup.classList.add('list-group', 'list-group-flush');
